requirements.txt
flask
vertexai
google-cloud-aiplatform
pillow